Added Following Changes: 1. Modularized image_gen.py into a more readable structure. 2. Removed all the redundant fonts fallbacks and imports into languages.py and centralized it. 3. Now Table, Charts, AI_Image,, text_utils everything has its own py file and image_gen.py is the main orchestration for all.
