+++ synthdoc/workflows/raw_document_generator/__init__.py
@@ -1,3 +1,29 @@
 from .workflow import RawDocumentGenerator
+from .image_gen import create_multicolumn_document_image
+from .chart_generator import create_advanced_chart
+from .table_generator import create_advanced_table
+from .ai_image_generator import (
+    generate_contextual_image,
+    create_placeholder_image,
+    generate_ai_image,
+)
+from .text_utils import (
+    wrap_text_to_width,
+    extract_keywords,
+    categorize_content,
+)
+from ...languages import load_language_font
 
-__all__ = ["RawDocumentGenerator"]
+__all__ = [
+    "RawDocumentGenerator",
+    "create_multicolumn_document_image",
+    "create_advanced_chart",
+    "create_advanced_table",
+    "generate_contextual_image",
+    "create_placeholder_image",
+    "generate_ai_image",
+    "wrap_text_to_width",
+    "extract_keywords",
+    "categorize_content",
+    "load_language_font",
+]
